LPS-115717 Rename for consistency
